Refactors RawData to use uint instead of ulong
Updates the RawData property in IOpenThermData interface and related classes to use uint (32-bit unsigned integer) instead of ulong (64-bit unsigned integer).

This change aligns the data type with the 32-bit OpenTherm frame standard, reducing memory usage and simplifying data handling.

+++ Drivers/DriverBase/Interfaces/IOpenThermData.cs
@@ -8,9 +8,9 @@ namespace TekuSP.Drivers.DriverBase.Interfaces
     public interface IOpenThermData
     {
         /// <summary>
-        /// Raw data
+        /// Raw data (32-bit OpenTherm frame)
         /// </summary>
-        ulong RawData
+        uint RawData
         {
             get; set;
         }
