+++ Drivers/Nano-OpenTherm/Requests/GetDHWBurnerOperationHoursRequest.cs
@@ -9,8 +9,8 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
 
         public ushort Hours { get; set; }
 
-        protected override ulong GetRawDataCore() => ProcessRequest(Hours);
-        protected override void SetRawDataCore(ulong value) { Hours = Utilities.GetLowUShort(value); }
+        protected override uint GetRawDataCore() => ProcessRequest(Hours);
+        protected override void SetRawDataCore(uint value) { Hours = Utilities.GetLowUShort(value); }
 
         public override MessageType MessageType => MessageType.READ_DATA;
         public override MessageID MessageID => MessageID.DHWBurnerOperationHours;
