+++ Drivers/Nano-OpenTherm/Requests/GetSerialRequest.cs
@@ -12,8 +12,8 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
         /// </summary>
         public byte Index { get; set; }
 
-        protected override ulong GetRawDataCore() => ProcessRequest(Index);
-        protected override void SetRawDataCore(ulong value)
+        protected override uint GetRawDataCore() => ProcessRequest(Index);
+        protected override void SetRawDataCore(uint value)
         {
             Index = Utilities.GetLowByte(value);
         }
