+++ Drivers/Nano-OpenTherm/Requests/SetCoolingControlRequest.cs
@@ -19,13 +19,13 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
             set => _percent = value.Normalize();
         }
 
-        protected override ulong GetRawDataCore()
+        protected override uint GetRawDataCore()
         {
             var raw = (uint)(Percent * 256f);
             return ProcessRequest(raw);
         }
 
-        protected override void SetRawDataCore(ulong value)
+        protected override void SetRawDataCore(uint value)
         {
             Percent = Utilities.GetFloat(value);
         }
